LUN-449: Fixes
